refactor: move `getMarketingCta` to schema/users
